add: custom hook to home screen
