+++ package.json
@@ -18,7 +18,8 @@
     "react-native": "0.71.6",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
-    "expo-linear-gradient": "~12.1.2"
+    "expo-linear-gradient": "~12.1.2",
+    "lottie-react-native": "5.1.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
